retrieve and delete managed item in same session

+++ src/main/java/io/michaeljgkopp/github/springconsoledemo/service/impl/CustomerServiceImpl.java
@@ -37,7 +37,10 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     @Transactional
-    public void deleteCustomer(Customer customer) {
-        customerDAO.delete(customer);
+    public void deleteCustomer(Long id) {
+        Customer customer = customerDAO.findById(id);
+        if (customer != null) {
+            customerDAO.delete(customer);
+        }
     }
 }
